Fix signed/unsigned mismatch warnings, part 8.
These f2s.c changes don't have any d2s.c counterparts.

Relevant signatures:

```
int32_t pow5bits(int32_t e);
uint32_t mulPow5InvDivPow2(uint32_t m, uint32_t q, int32_t j);
uint32_t mulPow5divPow2(uint32_t m, uint32_t i, int32_t j);
```

+++ ryu/f2s.c
@@ -199,8 +199,8 @@ static inline floating_decimal_32 f2d(const uint32_t ieeeMantissa, const uint32_
       // We need to know one removed digit even if we are not going to loop below. We could use
       // q = X - 1 above, except that would require 33 bits for the result, and we've found that
       // 32-bit arithmetic is faster even on 64-bit machines.
-      const int32_t l = FLOAT_POW5_INV_BITCOUNT + pow5bits(q - 1) - 1;
-      lastRemovedDigit = (uint8_t) (mulPow5InvDivPow2(mv, q - 1, -e2 + q - 1 + l) % 10);
+      const int32_t l = FLOAT_POW5_INV_BITCOUNT + pow5bits((int32_t) (q - 1)) - 1;
+      lastRemovedDigit = (uint8_t) (mulPow5InvDivPow2(mv, q - 1, -e2 + (int32_t) q - 1 + l) % 10);
     }
     if (q <= 9) {
       // The largest power of 5 that fits in 24 bits is 5^10, but q <= 9 seems to be safe as well.
@@ -219,17 +219,17 @@ static inline floating_decimal_32 f2d(const uint32_t ieeeMantissa, const uint32_
     const int32_t i = -e2 - (int32_t) q;
     const int32_t k = pow5bits(i) - FLOAT_POW5_BITCOUNT;
     int32_t j = (int32_t) q - k;
-    vr = mulPow5divPow2(mv, i, j);
-    vp = mulPow5divPow2(mp, i, j);
-    vm = mulPow5divPow2(mm, i, j);
+    vr = mulPow5divPow2(mv, (uint32_t) i, j);
+    vp = mulPow5divPow2(mp, (uint32_t) i, j);
+    vm = mulPow5divPow2(mm, (uint32_t) i, j);
 #ifdef RYU_DEBUG
     printf("%u * 5^%d / 10^%u\n", mv, -e2, q);
     printf("%u %d %d %d\n", q, i, k, j);
     printf("V+=%u\nV =%u\nV-=%u\n", vp, vr, vm);
 #endif
     if (q != 0 && (vp - 1) / 10 <= vm / 10) {
-      j = q - 1 - (pow5bits(i + 1) - FLOAT_POW5_BITCOUNT);
-      lastRemovedDigit = (uint8_t) (mulPow5divPow2(mv, i + 1, j) % 10);
+      j = (int32_t) q - 1 - (pow5bits(i + 1) - FLOAT_POW5_BITCOUNT);
+      lastRemovedDigit = (uint8_t) (mulPow5divPow2(mv, (uint32_t) (i + 1), j) % 10);
     }
     if (q <= 1) {
       // {vr,vp,vm} is trailing zeros if {mv,mp,mm} has at least q trailing 0 bits.
